The class are further implemented, class change and NLog was added.

+++ Common_eBike_Pruefstand/Luefter/ILuefter.cs
@@ -14,7 +14,7 @@ namespace Common_eBike_Pruefstand
 
 
         #region properties
-        float Value { get; }
+        byte Value { get; }
         #endregion
     }
 }
